Make all foreign imports safe
The problem with unsafe in this context is that
it can introduce deadlocks when two Haskell threads
use the library and one of them locks some rows in
an unfinished transaction, then the unsafe foreign
call in the other thread waits for a lock on those
rows. The foreign call never returns and (depending
on where the blocking Haskell thread runs, e.g. on the
same OS thread the other Haskell thread is bound to)
the thread holding the locks never gets a chance
to complete its transaction and release the locks.

This leads to blocking until MySQL's lock wait timeout
kicks in.