Remove obsolete `fieldDefault` from `data Field`
This field seems to be of little use and has been shown to be the cause
of hard to debug segfaults.

The not-so-recent documentation for libmysql at
https://dev.mysql.com/doc/c-api/5.7/en/c-api-data-structures.html
states about the `char *def`:

> This is set only if you use mysql_list_fields().

moreover, the documentation also states at
https://dev.mysql.com/doc/c-api/5.7/en/mysql-list-fields.html

> As of MySQL 5.7.11, mysql_list_fields() is deprecated and will be removed
> in a future version of MySQL. Instead, use mysql_query() to execute
> a SHOW COLUMNS statement.

and finally the Haskell binding doesn't appear to even bind the
`mysql_list_fields()` function in the first place. Moreover, the
mariadb-c-connector fork which has been supplanting `libmysql` in many
Linux distributions has been known to cause such segfaults.

Instead, the code incorrectly assumes the `char *def` field to be set
even when `mysql_list_fields()` was not used; and then it depends on
whether the MYSQL_FIELD struct happened to be allocated in zeroed out
memory region or not, on whether `char *def` and its associated
`def_length` field happen to be set to a zero or contain random garbage
resulting in either random HeapOverflow exceptions or straight out
memory access violations.

This patch simply removes the `fieldValue` field from `data Field` as
it appears to be of no use given the feature has long been effectively
obsoleted. This also requires a major version bump as per PVP.